+++ packages/soda-heroui/src/index.ts
@@ -1,3 +1,4 @@
+export * from "@/components/AutoRefresh"
 export * from "@/components/ErrorMessage"
 export * from "@/components/FormAutocomplete"
 export * from "@/components/FormCalendar"