+++ packages/soda-heroui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # soda-tanstack-form
 
+## 0.5.0
+
+### Minor Changes
+
+- 新增 AutoRefresh 相关方法和组件
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.8.0
+    - soda-type@6.2.0
+    - deepsea-tools@5.25.1
+
 ## 0.4.3
 
 ### Patch Changes