+++ packages/soda-type/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-type
 
+## 6.2.0
+
+### Minor Changes
+
+- 新增 AutoRefresh 相关方法和组件
+
 ## 6.1.1
 
 ### Patch Changes